requirements.txt
streamlit==1.32.0
python-pptx==0.6.23
python-dotenv==1.0.1
requests==2.31.0
reportlab==4.1.0 